@@ -12,11 +12,13 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@emotion/react": "^11.10.6",
     "@material-ui/core": "^4.12.4",
-    "@material-ui/data-grid": "^4.0.0-alpha.10",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@svgr/webpack": "^5.4.0",
+    "@material-ui/data-grid": "^4.0.0-alpha.37",
+    "@mui/icons-material": "^5.11.11",
+    "@mui/lab": "^5.0.0-alpha.121",
+    "@mui/material": "^5.11.11",
+    "@mui/styles": "^5.11.11",
     "@zeit/next-css": "^1.0.1",
     "@zeit/next-sass": "^1.0.1",
     "axios": "^0.21.1",
@@ -27,12 +29,13 @@
     "isomorphic-unfetch": "^3.0.0",
     "js-cookie": "^2.2.1",
     "jwt-decode": "^3.0.0",
-    "material-table": "^1.69.2",
+    "lodash": "^4.17.21",
+    "material-table": "^2.0.3",
     "material-ui-color-picker": "^3.5.0",
     "material-ui-dropzone": "^3.3.1",
     "moment": "^2.27.0",
     "moment-timezone": "^0.5.31",
-    "next": "^11.1.3",
+    "next": "^12.3.4",
     "next-i18next": "^8.1.3",
     "next-redux-wrapper": "^6.0.2",
     "papaparse": "^5.3.0",
@@ -40,15 +43,15 @@
     "postcss": "^8.2.15",
     "postcss-selector-parser": "^6.0.6",
     "prop-types": "^15.7.2",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^1.0.1",
     "react-device-detect": "^1.17.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-google-recaptcha": "^2.1.0",
     "react-gtm-module": "^2.0.11",
     "react-https-redirect": "^1.1.0",
     "react-image-gallery": "^1.0.8",
-    "react-material-ui-carousel": "^2.1.1",
+    "react-material-ui-carousel": "^3.4.2",
     "react-material-ui-form-validator": "^2.1.1",
     "react-moment": "^0.9.7",
     "react-number-format": "^4.4.1",
@@ -66,7 +69,9 @@
   },
   "devDependencies": {
     "@babel/preset-env": "^7.14.4",
-    "babel-plugin-styled-components": "^2.0.2"
+    "@svgr/webpack": "^6.5.1",
+    "babel-plugin-styled-components": "^2.0.2",
+    "url-loader": "^4.1.1"
   },
   "engines": {
     "node": "14.x",
