@@ -18,14 +18,14 @@ import Wishlist from 'public/images/svg/wishlist.svg';
 import AddressBook from 'public/images/svg/notebook.svg';
 import Shield from 'public/images/svg/shield.svg';
 import CarWheel from 'public/images/svg/car_wheel.svg';
-import DeleteOutlinedIcon from '@material-ui/icons/DeleteOutlined';
+import { DeleteIcon as DeleteOutlinedIcon } from '@mui/icons-material/Delete';
 import Delivery from 'public/images/svg/delivery.svg';
 import Paypal from 'public/images/svg/cc-pay.svg';
 import Mastercard from 'public/images/svg/cc-master.svg';
 import Visa from 'public/images/svg/cc-visa.svg';
 import Discovery from 'public/images/svg/cc-discover.svg';
 import Cart from 'public/images/svg/cart.svg';
-import ShoppingCartOutlinedIcon from '@material-ui/icons/ShoppingCartOutlined';
+import { ShoppingCartIcon as ShoppingCartOutlinedIcon } from '@mui/icons-material/ShoppingCart';
 import Twitter from 'public/images/svg/social-twitter.svg';
 import Facebook from 'public/images/svg/social-facebook.svg';
 import Pininterest from 'public/images/svg/social-pinterest.svg';
@@ -41,7 +41,7 @@ import Computer from 'public/images/svg/categories/computer.svg';
 import Furniture from 'public/images/svg/categories/furniture.svg';
 import Sport from 'public/images/svg/categories/sport.svg';
 import Baby from 'public/images/svg/categories/baby.svg';
-import AddIcon from '@material-ui/icons/Add';
+import AddIcon from '@mui/icons-material/Add';
 import Car from 'public/images/svg/categories/car.svg';
 import Garden from 'public/images/svg/categories/garden.svg';
 import Health from 'public/images/svg/categories/health.svg';
@@ -53,26 +53,26 @@ import Category from 'public/images/svg/category.svg';
 import Vendor from 'public/images/svg/vendor.svg';
 import Brand from 'public/images/svg/brand.svg';
 import Supplies from 'public/images/svg/supplies.svg';
-import FavoriteBorderIcon from '@material-ui/icons/FavoriteBorder';
+import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
 import LogoIconWhite from 'public/images/logoIconWhite.svg';
-import PersonAddOutlinedIcon from '@material-ui/icons/PersonAddOutlined';
-import StoreMallDirectoryOutlinedIcon from '@material-ui/icons/StoreMallDirectoryOutlined';
-import FavoriteIcon from '@material-ui/icons/Favorite';
-import ImageSearchIcon from '@material-ui/icons/ImageSearch';
-import SearchIcon from '@material-ui/icons/Search';
-import CloseIcon from '@material-ui/icons/Close';
-import StorefrontIcon from '@material-ui/icons/Storefront';
-import ExitToAppIcon from '@material-ui/icons/ExitToApp';
-import ImportExportIcon from '@material-ui/icons/ImportExport';
-import ArrowBackIcon from '@material-ui/icons/ArrowBack';
-import HomeIcon from '@material-ui/icons/Home';
-import AddCircleIcon from '@material-ui/icons/AddCircle';
-import FormatListBulletedIcon from '@material-ui/icons/FormatListBulleted';
-import PermIdentityOutlinedIcon from '@material-ui/icons/PermIdentityOutlined';
-import PrintIcon from '@material-ui/icons/Print';
-import EditIcon from '@material-ui/icons/Edit';
-import KeyboardBackspaceIcon from '@material-ui/icons/KeyboardBackspace';
-import ColorLensIcon from '@material-ui/icons/ColorLens';
+import PersonAddOutlinedIcon from '@mui/icons-material/PersonAddOutlined';
+import StoreMallDirectoryOutlinedIcon from '@mui/icons-material/StoreMallDirectoryOutlined';
+import FavoriteIcon from '@mui/icons-material/Favorite';
+import ImageSearchIcon from '@mui/icons-material/ImageSearch';
+import SearchIcon from '@mui/icons-material/Search';
+import CloseIcon from '@mui/icons-material/Close';
+import StorefrontIcon from '@mui/icons-material/Storefront';
+import ExitToAppIcon from '@mui/icons-material/ExitToApp';
+import ImportExportIcon from '@mui/icons-material/ImportExport';
+import ArrowBackIcon from '@mui/icons-material/ArrowBack';
+import HomeIcon from '@mui/icons-material/Home';
+import AddCircleIcon from '@mui/icons-material/AddCircle';
+import FormatListBulletedIcon from '@mui/icons-material/FormatListBulleted';
+import PermIdentityOutlinedIcon from '@mui/icons-material/PermIdentityOutlined';
+import PrintIcon from '@mui/icons-material/Print';
+import EditIcon from '@mui/icons-material/Edit';
+import KeyboardBackspaceIcon from '@mui/icons-material/KeyboardBackspace';
+import ColorLensIcon from '@mui/icons-material/ColorLens';
 import Bag from 'public/images/svg/categories/bag.svg';
 import CloseBag from 'public/images/svg/categories/closebag.svg';
 import Ballons from 'public/images/svg/categories/ballons.svg';
@@ -91,21 +91,21 @@ import Fiestas from 'public/images/svg/categories/fiestas.svg';
 import Envases from 'public/images/svg/categories/envases.svg';
 import Covid from 'public/images/svg/categories/covid.svg';
 import Accesorios from 'public/images/svg/categories/accesorios.svg';
-import PaymentIcon from '@material-ui/icons/Payment';
-import AttachMoneyIcon from '@material-ui/icons/AttachMoney';
-import CheckCircleOutlineIcon from '@material-ui/icons/CheckCircleOutline';
-import ErrorOutlineIcon from '@material-ui/icons/ErrorOutline';
+import PaymentIcon from '@mui/icons-material/Payment';
+import AttachMoneyIcon from '@mui/icons-material/AttachMoney';
+import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
+import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline';
 import Christmas from 'public/images/svg/categories/christmas.svg';
 import Others from 'public/images/svg/categories/others.svg';
 import NoImages from 'public/images/svg/noimage.svg';
 import Coupon from 'public/images/svg/coupon.svg';
 import Blackfriday from 'public/images/svg/categories/blackfriday.svg';
-import SentimentVeryDissatisfiedIcon from '@material-ui/icons/SentimentVeryDissatisfied';
-import ExtensionIcon from '@material-ui/icons/ExtensionOutlined';
+import SentimentVeryDissatisfiedIcon from '@mui/icons-material/SentimentVeryDissatisfied';
+import ExtensionIcon from '@mui/icons-material/Extension';
 import Filter from 'public/images/svg/filter.svg';
 import Typography from './Typography';
-import SettingsIcon from '@material-ui/icons/Settings';
-import DoneIcon from '@material-ui/icons/Done';
+import SettingsIcon from '@mui/icons-material/Settings';
+import DoneIcon from '@mui/icons-material/Done';
 
 
 const styles = (theme) => ({
